CWEval/requirements/core.txt
fire
natsort
pytest
pytest-xdist
wrapt_timeout_decorator
pytest-timeout
requests
lxml
pyyaml
PyJWT
argon2-cffi
pycryptodome
docker
numpy
p_tqdm
